Big change for bikeshed, auto get mod lists, test logs, test profiles. etc.

+++ report.sh
@@ -1,8 +1,5 @@
 #! /bin/sh -e
 
-if [ ! -f tests-out ]; then
-    exit 1
-fi
 
 toaddr="dev.null@example.com"
 fromaddr="Auto CI <bot@example.com>"
@@ -16,6 +13,12 @@ dock="sudo docker"
 dockimage="$dock images $image"
 dockrun="sudo docker run --rm -it $image"
 
+tests="test-$(cat latest-Fedora-Modular-27.COMPOSE_ID)"
+otests="test-$(cat prev-Fedora-Modular-27.COMPOSE_ID)"
+if [ ! -d $tests ]; then
+    exit 1
+fi
+
 rm -f /tmp/out.$$
 echo " Auto CI for image jamesantill/boltron-27:" > /tmp/out.$$
 $dockimage:latest --format='ID: {{.ID}}' --no-trunc >> /tmp/out.$$
@@ -27,83 +30,96 @@ echo "    Base   : $(cat latest-Fedora-Modular-27.COMPOSE_ID)" >> /tmp/out.$$
 if [ -f prev-Fedora-Modular-27.COMPOSE_ID ]; then
 echo "    Prev   : $(cat prev-Fedora-Modular-27.COMPOSE_ID)" >> /tmp/out.$$
 fi
+if false; then
 echo "    Bike   : $($dockrun cat latest-Fedora-Modular-Bikeshed.COMPOSE_ID)" >> /tmp/out.$$
 echo "    Prev   : $($dockrun:$(cat prev-Fedora-Modular-27.COMPOSE_ID) cat latest-Fedora-Modular-Bikeshed.COMPOSE_ID)" >> /tmp/out.$$
+fi
 
 echo "" >> /tmp/out.$$
 
 echo "==================================================" >> /tmp/out.$$
 printf "%30s\n" "Summary" >> /tmp/out.$$
 echo "--------------------------------------------------" >> /tmp/out.$$
-tm=$(perl -MFile::stat -le '$num = stat("tests-end")->mtime - stat("tests-beg")->mtime; print int($num / 60), ":", $num % 60')
-echo " Time taken: $tm" >> /tmp/out.$$
-echo " Modules   : $(wc -l < tests-out)" >> /tmp/out.$$
-echo "    PASS: $(fgrep pass tests-out | wc -l)" >> /tmp/out.$$
-echo "    FAIL: $(fgrep FAIL tests-out | wc -l)" >> /tmp/out.$$
-echo "        FAIL dnf err : $(fgrep 'FAIL: DNF' tests-out | wc -l)" >> /tmp/out.$$
-echo "        FAIL dnf skip: $(fgrep 'FAIL: Did' tests-out | wc -l)" >> /tmp/out.$$
+tm=$(perl -MFile::stat -le '$num = stat($ARGV[0] . ".STATUS")->mtime - stat($ARGV[0] . ".COMPOSE_ID")->mtime; $h = int($num / 60); if ($h >= 60) { printf "%2d:%02d:%02d", int($h / 60), $h % 60, $num % 60 } else { printf "   %d:%02d", $h, $num % 60 }' latest-Fedora-Modular-27)
+echo " Compose taken: $tm" >> /tmp/out.$$
+tm=$(perl -MFile::stat -le '$num = stat($ARGV[0] . "/end")->mtime - stat($ARGV[0] . "/beg")->mtime; $h = int($num / 60); if ($h >= 60) { printf "%2d:%02d:%02d", int($h / 60), $h % 60, $num % 60 } else { printf "   %d:%02d", $h, $num % 60 }' $tests)
+echo " Tests taken  : $tm" >> /tmp/out.$$
+echo " Modules      : $(wc -l < $tests/mods)" >> /tmp/out.$$
+echo " Tests        : $(wc -l < $tests/out)" >> /tmp/out.$$
+echo "    PASS: $(fgrep pass $tests/out | wc -l)" >> /tmp/out.$$
+echo "    FAIL: $(fgrep FAIL $tests/out | wc -l)" >> /tmp/out.$$
+echo "        FAIL dnf err : $(fgrep 'FAIL: DNF' $tests/out | wc -l)" >> /tmp/out.$$
+echo "        FAIL dnf skip: $(fgrep 'FAIL: Did' $tests/out | wc -l)" >> /tmp/out.$$
 echo "" >> /tmp/out.$$
 
 if [ -f prev-Fedora-Modular-27.COMPOSE_ID ]; then
-diffo="$(cat prev-Fedora-Modular-27.COMPOSE_ID)-list-modules"
-diffn="$(cat latest-Fedora-Modular-27.COMPOSE_ID)-list-modules"
-sudo docker run --rm -v $(pwd):/mnt  jamesantill/boltron-27:$(cat prev-Fedora-Modular-27.COMPOSE_ID) /mnt/list-modules-py3.py > $diffo
-sudo docker run --rm -v $(pwd):/mnt  jamesantill/boltron-27:$(cat latest-Fedora-Modular-27.COMPOSE_ID) /mnt/list-modules-py3.py > $diffn
-if ! diff -u $diffo $diffn > /dev/null; then
+if ! diff -u $otests/mods $tests/mods > /dev/null; then
 echo "==================================================" >> /tmp/out.$$
 printf "%30s\n" "Modules diff" >> /tmp/out.$$
 echo "--------------------------------------------------" >> /tmp/out.$$
-diff -u $diffo $diffn >> /tmp/out.$$ || true
+diff -u $otests/mods $tests/mods >> /tmp/out.$$ || true
 fi
-rm -f $diffo $diffn
 
-diffo="$(cat prev-Fedora-Modular-27.COMPOSE_ID)-rpm"
-diffn="$(cat latest-Fedora-Modular-27.COMPOSE_ID)-rpm"
-sudo docker run --rm -v $(pwd):/mnt  jamesantill/boltron-27:$(cat prev-Fedora-Modular-27.COMPOSE_ID) /mnt/list-rpm.sh > $diffo
-sudo docker run --rm -v $(pwd):/mnt  jamesantill/boltron-27:$(cat prev-Fedora-Modular-27.COMPOSE_ID) /mnt/list-rpm.sh > $diffn
-
-if ! diff -u $diffo $diffn > /dev/null; then
+if ! diff -u $otests/rpm $tests/rpm > /dev/null; then
 echo "==================================================" >> /tmp/out.$$
 printf "%30s\n" "RPM change diff" >> /tmp/out.$$
 echo "--------------------------------------------------" >> /tmp/out.$$
-diff -u $diffo $diffn >> /tmp/out.$$ || true
+diff -u $otests/rpm $tests/rpm >> /tmp/out.$$ || true
 fi
-rm -f $diffo $diffn
 
 else
 # No previous image ...
 echo "==================================================" >> /tmp/out.$$
 printf "%30s\n" "Modules" >> /tmp/out.$$
 echo "--------------------------------------------------" >> /tmp/out.$$
-sudo docker run --rm -v $(pwd):/mnt  jamesantill/boltron-27:$(cat prev-Fedora-Modular-27.COMPOSE_ID) /mnt/list-modules-py3.py >> /tmp/out.$$
+cat $tests/mods >> /tmp/out.$$ || true
 fi
 echo "" >> /tmp/out.$$
 
 echo "" >> /tmp/out.$$
 echo " Full output:" >> /tmp/out.$$
-cat tests-out >> /tmp/out.$$
+cat $tests/out >> /tmp/out.$$
 echo "" >> /tmp/out.$$
 echo "" >> /tmp/out.$$
 echo " Full image used data:" >> /tmp/out.$$
-cat tests-hdr >> /tmp/out.$$
+cat $tests/hdr >> /tmp/out.$$
 
 perl -i -pe 's/\r//g' /tmp/out.$$
 
 if [ "x$1" = "xemail" ]; then
-if [ "x$(fgrep pass tests-out | wc -l)" != "x$(wc -l < tests-out)" ]; then
-mail -n -r "$fromaddr" -s "Modularity Image $(sudo docker images jamesantill/boltron-27:latest --format='ID: {{.ID}}') FAIL: $(fgrep FAIL tests-out | wc -l)/$(wc -l < tests-out)" "$toaddr" < /tmp/out.$$
+if [ "x$(fgrep pass $tests/out | wc -l)" != "x$(wc -l < $tests/out)" ]; then
+mail -n -r "$fromaddr" -s "Modularity Image $(sudo docker images jamesantill/boltron-27:latest --format='ID: {{.ID}}') FAIL: $(fgrep FAIL $tests/out | wc -l)/$(wc -l < $tests/out)" "$toaddr" < /tmp/out.$$
 fi
 rm -f /tmp/out.$$
 exit 0
 fi
+
+for i in $(fgrep 'FAIL: DNF' $tests/out | awk '{ print $1 }'); do
+    j="$(echo $i | tr '/' '-')"
+echo "==================================================" >> /tmp/out.$$
+printf "%30s\n" "Test output: $i" >> /tmp/out.$$
+echo "--------------------------------------------------" >> /tmp/out.$$
+if [ -s $tests/out-$j-1 ]; then
+echo "--------------------------------------------------" >> /tmp/out.$$
+printf "%30s\n" "stdout" >> /tmp/out.$$
+echo "--------------------------------------------------" >> /tmp/out.$$
+cat $tests/out-$j-1  >> /tmp/out.$$
+fi
+if [ -s $tests/out-$j-2 ]; then
+echo "--------------------------------------------------" >> /tmp/out.$$
+printf "%30s\n" "STDERR" >> /tmp/out.$$
+echo "--------------------------------------------------" >> /tmp/out.$$
+cat $tests/out-$j-2  >> /tmp/out.$$
+fi
+done
 
 if [ "x$1" = "xemail-force" ]; then
-mail -n -r "$fromaddr" -s "Modularity Image $(sudo docker images jamesantill/boltron-27:latest --format='ID: {{.ID}}') FAIL: $(fgrep FAIL tests-out | wc -l)/$(wc -l < tests-out)" "$toaddr" < /tmp/out.$$
+mail -n -r "$fromaddr" -s "Modularity Image $(sudo docker images jamesantill/boltron-27:latest --format='ID: {{.ID}}') FAIL: $(fgrep FAIL $tests/out | wc -l)/$(wc -l < $tests/out)" "$toaddr" < /tmp/out.$$
 rm -f /tmp/out.$$
 exit 0
 fi
 
-echo "Subject: Modularity Image $(sudo docker images jamesantill/boltron-27:latest --format='ID: {{.ID}}') FAIL: $(fgrep FAIL tests-out | wc -l)/$(wc -l < tests-out)"
+echo "Subject: Modularity Image $(sudo docker images jamesantill/boltron-27:latest --format='ID: {{.ID}}') FAIL: $(fgrep FAIL $tests/out | wc -l)/$(wc -l < $tests/out)"
 
 cat /tmp/out.$$
